+++ src/main/java/net/crusadergames/bugwars/Util/Constants.java
@@ -4,4 +4,18 @@ public class Constants {
 
     public static final String RESPONSE_SCRIPTDELETED = "Script Deleted";
 
+    public static final String CONDITIONALCOMMAND_IFENEMY = "ifEnemy";
+    public static final String CONDITIONALCOMMAND_IFALLY = "ifAlly";
+    public static final String CONDITIONALCOMMAND_IFFOOD = "ifFood";
+    public static final String CONDITIONALCOMMAND_IFEMPTY = "ifEmpty";
+    public static final String CONDITIONALCOMMAND_IFWALL = "ifWall";
+    public static final String CONDITIONALCOMMAND_GOTO = "goto";
+
+    public static final String ACTIONCOMMAND_NOOP = "noop";
+    public static final String ACTIONCOMMAND_MOV = "mov";
+    public static final String ACTIONCOMMAND_ROTR = "rotr";
+    public static final String ACTIONCOMMAND_ROTL = "rotl";
+    public static final String ACTIONCOMMAND_ATT = "att";
+    public static final String ACTIONCOMMAND_EAT = "eat";
+
 }
