+++ src/test/java/net/crusadergames/bugwars/service/ParserServiceTest.java
@@ -1,0 +1,12 @@
+package net.crusadergames.bugwars.service;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class ParserServiceTest {
+
+    @Test
+    void returnByteCode() {
+    }
+}
